docs(miniapps): store-surface guide + react-refresh split for the artwork module
- developer guide gains the spec-077 store-surface section: host-curated
  artwork rules, badge honesty, and the ?view= sub-view seam
- appArtwork splits into appArt.jsx (components only) + appArtwork.js
  (slug map + artworkFor), the scopeGate/scopeControls pattern, clearing
  react-refresh/only-export-components; consumers are unchanged
- tasks.md checked off through T019
